requirements.txt
# Minimum Django and REST framework version
Django>=1.8
djangorestframework>=3.1

# Test requirements
pytest-django
pytest
pytest-cov
flake8
django-model-utils

# wheel for PyPI installs
#wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.15.3
